+++ src/test/kotlin/com/tezos/lang/michelson/editor/completion/MichelsonComplexTypeCompletionContributorTest.kt
@@ -7,7 +7,7 @@ import com.tezos.lang.michelson.lang.MichelsonLanguage
  */
 class MichelsonComplexTypeCompletionContributorTest:MichelsonCompletionTest(){
     fun testCompletion() {
-        val referenceComplex = MichelsonLanguage.TYPES_COMPLEX.toTypedArray()
+        val referenceComplex = MichelsonLanguage.TYPES_NESTED.toTypedArray()
         val referenceComparable = MichelsonLanguage.TYPES_COMPARABLE.toTypedArray()
         val all = referenceComplex + referenceComparable